Move login fetch command to merlins connector
