+++ web/app/containers/login/reducer.js
@@ -1,7 +1,7 @@
 import Immutable from 'immutable'
 import {handleActions} from 'redux-actions'
 import {mergePayloadForActions} from '../../utils/reducer-utils'
-import {receiveData} from './actions'
+import {receiveLoginPageData} from '../../integration-manager/login/responses'
 
 const signinFields = [
     {
@@ -117,5 +117,5 @@ const initialState = Immutable.fromJS({
 })
 
 export default handleActions({
-    ...mergePayloadForActions(receiveData)
+    ...mergePayloadForActions(receiveLoginPageData)
 }, initialState)
