+++ web/app/integration-manager/login/responses.js
@@ -1,0 +1,3 @@
+import {createAction} from '../../utils/utils'
+
+export const receiveLoginPageData = createAction('Receive  Login Page data')
